refactor data models
