+++ frontend/AngularApp/src/app/shared/models/data-to-api/user.ts
@@ -5,3 +5,8 @@ export type UserRegister = {
   email: string;
   password: string;
 };
+
+export type UserLogin = {
+  username: string;
+  password: string;
+};
